refactor(oh7): move Provider trait into core (dependency inversion) (#252)
Move providers::traits -> core::provider (Provider trait, CompletionRequest/
Response, ChatMessage, ProviderMetadata, TokenStream). The orchestration
engine (core) calls this contract; providers now implements it and depends
on core, not the reverse. Repoints ~20 import sites across core/orchestration/
es/*, providers/*, providers/api/*, and cli/{audit,run}.rs.

ProviderMetadata's fields are constructed by every Provider::metadata() impl
but only read in tests; this was previously silent because `mod providers;`
in main.rs carries a blanket #[allow(dead_code)], which core lacks. Scoped
#[allow(dead_code)] added directly on the struct to preserve the exact prior
suppression (no behavior change) rather than widening core's allow.
